Share the step configurator href logic between the SelectConnection and AddStep components
