+++ app/ui-react/packages/api/tests/helpers/integrationFunctions.spec.tsx
@@ -1,4 +1,4 @@
-import { getEmptyIntegration, getStepIcon } from '../../src';
+import { getEmptyIntegration, getIntegrationStepIcon } from '../../src';
 import { IConnectionWithIconFile, Integration, Step } from '@syndesis/models';
 import expect = require('expect');
 
@@ -27,7 +27,7 @@ export default describe('integrationFunctions', () => {
         },
       ],
     } as Integration;
-    const iconPath = getStepIcon('', integration, 'id', 0);
+    const iconPath = getIntegrationStepIcon('', integration, 'id', 0);
     expect(iconPath).toBe('data:blah');
   });
   it('Should return a valid icon URL for a legacy connection', () => {
@@ -46,7 +46,7 @@ export default describe('integrationFunctions', () => {
         },
       ],
     } as Integration;
-    const iconPath = getStepIcon('', integration, 'id', 0);
+    const iconPath = getIntegrationStepIcon('', integration, 'id', 0);
     expect(iconPath).toBe('./../../icons/blah.connection.png');
   });
   it('Should return a valid icon URL for a connection that references a db entity', () => {
@@ -66,7 +66,7 @@ export default describe('integrationFunctions', () => {
         },
       ],
     } as Integration;
-    const iconPath = getStepIcon('', integration, 'id', 0);
+    const iconPath = getIntegrationStepIcon('', integration, 'id', 0);
     expect(iconPath).toBe('/connectors/foo/icon?db:blah');
   });
   it('Should return a valid icon URL for a connection that references a db entity', () => {
@@ -86,7 +86,7 @@ export default describe('integrationFunctions', () => {
         },
       ],
     } as Integration;
-    const iconPath = getStepIcon('', integration, 'id', 0);
+    const iconPath = getIntegrationStepIcon('', integration, 'id', 0);
     expect(iconPath).toBe('/connectors/foo/icon?db:blah');
   });
   it('Should return a valid icon URL for a connection that references a db entity', () => {
@@ -106,7 +106,7 @@ export default describe('integrationFunctions', () => {
         },
       ],
     } as Integration;
-    const iconPath = getStepIcon('', integration, 'id', 0);
+    const iconPath = getIntegrationStepIcon('', integration, 'id', 0);
     expect(iconPath).toBe('/connectors/foo/icon?extension:blah');
   });
   it('Should return a valid icon URL for a connection with an icon file', () => {
@@ -126,7 +126,7 @@ export default describe('integrationFunctions', () => {
         },
       ],
     } as Integration;
-    const iconPath = getStepIcon('', integration, 'id', 0);
+    const iconPath = getIntegrationStepIcon('', integration, 'id', 0);
     expect(iconPath).toBe('foo');
   });
   it('Should return a valid icon URL for an extension step', () => {
@@ -145,7 +145,7 @@ export default describe('integrationFunctions', () => {
         },
       ],
     } as Integration;
-    const iconPath = getStepIcon('', integration, 'id', 0);
+    const iconPath = getIntegrationStepIcon('', integration, 'id', 0);
     expect(iconPath).toBe('blah');
   });
   it('Should return a valid icon URL for a step', () => {
@@ -162,7 +162,7 @@ export default describe('integrationFunctions', () => {
         },
       ],
     } as Integration;
-    const iconPath = getStepIcon('', integration, 'id', 0);
+    const iconPath = getIntegrationStepIcon('', integration, 'id', 0);
     expect(iconPath).toBe('/icons/steps/log.svg');
   });
 });
